Move file validation into utils folder
