+++ src/applications/burials-ez/config/chapters/03-military-history/uploadDD214.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { fileInputMultipleSchema } from '~/platform/forms-system/src/js/web-component-patterns';
 import { burialUploadUI } from '../../../utils/upload';
 import { generateTitle } from '../../../utils/helpers';
+import { validateFileUploads } from '../../../utils/validation';
 
 export default {
   uiSchema: {
@@ -22,42 +23,10 @@ export default {
       </>
     ),
     militarySeparationDocuments: {
-      ...burialUploadUI('Upload DD214 or other separation documents'),
-      'ui:validations': [
-        // Temporary workaround to enforce required file until bug is fixed
-        // Temporary workaround to ensure invalid files are not valid to continue
-        // https://github.com/department-of-veterans-affairs/vets-design-system-documentation/issues/4716
-        (errors, fieldData /* files array or single file */) => {
-          let files = [];
-
-          if (Array.isArray(fieldData)) {
-            files = fieldData;
-          } else if (fieldData) {
-            files = [fieldData];
-          }
-
-          // Required check: no files at all
-          if (files.length === 0) {
-            errors.addError('Upload a supporting document');
-            return;
-          }
-
-          files.forEach(file => {
-            if (file?.isEncrypted && !file?.confirmationCode) {
-              return;
-            }
-
-            if (!file || !file.name) {
-              errors.addError('Upload a supporting document');
-              return;
-            }
-
-            if (file.errorMessage) {
-              errors.addError(file.errorMessage);
-            }
-          });
-        },
-      ],
+      ...burialUploadUI({
+        title: 'Upload DD214 or other separation documents',
+      }),
+      'ui:validations': [validateFileUploads()],
     },
   },
   schema: {
