+++ src/applications/burials-ez/config/chapters/05-additional-information/additionalEvidence.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { fileInputMultipleSchema } from '~/platform/forms-system/src/js/web-component-patterns';
 import { burialUploadUI } from '../../../utils/upload';
+import { validateFileUploads } from '../../../utils/validation';
 import { generateTitle } from '../../../utils/helpers';
 
 export default {
@@ -102,35 +103,11 @@ export default {
       </>
     ),
     additionalEvidence: {
-      ...burialUploadUI('Upload additional supporting documents', false),
-      'ui:validations': [
-        // Temporary workaround to ensure invalid files are not valid to continue
-        // https://github.com/department-of-veterans-affairs/vets-design-system-documentation/issues/4716
-        (errors, fieldData /* files array or single file */) => {
-          let files = [];
-
-          if (Array.isArray(fieldData)) {
-            files = fieldData;
-          } else if (fieldData) {
-            files = [fieldData];
-          }
-
-          files.forEach(file => {
-            if (file?.isEncrypted && !file?.confirmationCode) {
-              return;
-            }
-
-            if (!file || !file.name) {
-              errors.addError('Upload a supporting document');
-              return;
-            }
-
-            if (file.errorMessage) {
-              errors.addError(file.errorMessage);
-            }
-          });
-        },
-      ],
+      ...burialUploadUI({
+        title: 'Upload additional supporting documents',
+        required: false,
+      }),
+      'ui:validations': [validateFileUploads({ required: false })],
     },
   },
   schema: {
