+++ src/applications/burials-ez/config/chapters/05-additional-information/transportationReceipts.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { fileInputMultipleSchema } from '~/platform/forms-system/src/js/web-component-patterns';
 import { burialUploadUI } from '../../../utils/upload';
+import { validateFileUploads } from '../../../utils/validation';
 
 import { generateTitle } from '../../../utils/helpers';
 
@@ -32,42 +33,8 @@ export default {
       </>
     ),
     transportationReceipts: {
-      ...burialUploadUI('Upload an itemized receipt'),
-      'ui:validations': [
-        // Temporary workaround to enforce required file until bug is fixed
-        // Temporary workaround to ensure invalid files are not valid
-        // https://github.com/department-of-veterans-affairs/vets-design-system-documentation/issues/4716
-        (errors, fieldData /* files array or single file */) => {
-          let files = [];
-
-          if (Array.isArray(fieldData)) {
-            files = fieldData;
-          } else if (fieldData) {
-            files = [fieldData];
-          }
-
-          // Required check: no files at all
-          if (files.length === 0) {
-            errors.addError('Upload a supporting document');
-            return;
-          }
-
-          files.forEach(file => {
-            if (file?.isEncrypted && !file?.confirmationCode) {
-              return;
-            }
-
-            if (!file || !file.name) {
-              errors.addError('Upload a supporting document');
-              return;
-            }
-
-            if (file.errorMessage) {
-              errors.addError(file.errorMessage);
-            }
-          });
-        },
-      ],
+      ...burialUploadUI({ title: 'Upload an itemized receipt' }),
+      'ui:validations': [validateFileUploads()],
     },
   },
   schema: {
